scripts/requirements.txt
google-cloud-bigquery>=2.34.4
google-cloud-storage>=2.7.0
